+++ src/django/api/helpers.py
@@ -3,6 +3,8 @@ import csv
 import json
 from unidecode import unidecode
 
+from api.constants import NumberOfWorkersRanges
+
 CONSONANT_SOUND = re.compile(r'''
 one(![ir])
 ''', re.IGNORECASE | re.VERBOSE)
@@ -82,3 +84,23 @@ def clean(column):
     if not column:
         column = None
     return column
+
+
+def value_is_in_range(range_value, range):
+    range_min = range.get('min', 0)
+    range_max = range.get('max', None)
+    value_min = range_value.get('min', 0)
+    value_max = range_value.get('max', value_min)
+    if value_min < range_min:
+        return value_max >= range_min
+    if value_min >= range_min:
+        return range_max is None or value_min <= range_max
+
+
+def convert_to_standard_ranges(number_of_workers, overlapping_ranges=None):
+    if overlapping_ranges is None:
+        overlapping_ranges = set()
+    for range in NumberOfWorkersRanges.STANDARD_RANGES:
+        if value_is_in_range(number_of_workers, range):
+            overlapping_ranges.add(range.get('label', ''))
+    return overlapping_ranges